+++ Conselho/Controllers/GradeSubject.php
@@ -60,6 +60,18 @@ class GradeSubject extends Controller
         return count(array_unique($school_ids)) > 1;
     }
 
+    private function check_permission(int $id = null) : bool {
+        $atlas = $this->atlas();
+
+        if ($id) {
+            $school_id = $atlas->fetchRecord($this->mapper_class_name, $id, ['grade'])->grade->school_id;
+        } else {
+            $school_id = $atlas->fetchRecord(GradeMapper::class, $this->input_int('grade_id'))->school_id;
+        }
+
+        return $this->has_permission('grade', $school_id);
+    }
+
     // METHODS
 
     public function get() : ?string {
@@ -93,6 +105,11 @@ class GradeSubject extends Controller
             return null;
         }
 
+        if (!$this->check_permission()) {
+            http_response_code(403);
+            return null;
+        }
+
         $data = $this->get_post_data();
         if (!$record = $this->insert($data)) {
             http_response_code(500);
@@ -106,6 +123,11 @@ class GradeSubject extends Controller
         if (!$record = $this->fetch($id)) {
             http_response_code(404);
             return;
+        }
+
+        if (!$this->check_permission($id)) {
+            http_response_code(403);
+            return null;
         }
 
         if (!$this->delete_with_dependencies($record)) {
